Refactored header/implementation files into include and src respectivley
