+++ beast/CMakeLists.txt
@@ -2,17 +2,17 @@ cmake_minimum_required(VERSION 3.10)
 project(beast-service)
 
 file (GLOB APP_FILES
-  beast.hpp
-  http_session.cpp
-  http_session.hpp
-  listener.cpp
-  listener.hpp
+  include/beast.hpp
+  src/http_session.cpp
+  include/http_session.hpp
+  src/listener.cpp
+  include/listener.hpp
   main.cpp
-  net.hpp
-  shared_state.cpp
-  shared_state.hpp
-  websocket_session.cpp
-  websocket_session.hpp
+  include/net.hpp
+  src/shared_state.cpp
+  include/shared_state.hpp
+  src/websocket_session.cpp
+  include/websocket_session.hpp
 )
 
 source_group ("" FILES ${APP_FILES})
