+++ beast/include/shared_state.hpp
