+++ beast/src/shared_state.cpp
@@ -1,5 +1,5 @@
-#include "shared_state.hpp"
-#include "websocket_session.hpp"
+#include "../include/shared_state.hpp"
+#include "../include/websocket_session.hpp"
 
 shared_state::
 shared_state(std::string doc_root)
